move handleSeacrhLogic to CodeListPage which already knows of code lists
